[Refactor] 공통부분들 묶기
