+++ banchan/src/components/utils/Price.jsx
@@ -1,5 +1,6 @@
 import styled from "styled-components";
-import theme from "./theme";
+import theme from "../utils/styles/theme.js";
+import { CenterContainer } from "./styles/common.jsx";
 
 const Price = ({ product }) => {
   const isOnSale = !!product.n_price;
@@ -14,9 +15,8 @@ const Price = ({ product }) => {
   );
 };
 
-const StyledPriceWrapper = styled.div`
-  display: flex;
-  align-items: center;
+const StyledPriceWrapper = styled(CenterContainer)`
+  justify-content: start;
 `;
 
 const StyledFinalPriceSpan = styled.span`
