+++ banchan/src/components/utils/button/TextButton.jsx
@@ -1,6 +1,7 @@
 import styled from "styled-components";
-import theme from "../constant";
-import { BUTTON_TYPE, Button } from "./Button";
+import theme from "../styles/theme";
+import { BUTTON_TYPE } from "../variables.js";
+import { Button } from "../styles/common";
 
 const TextButton = ({ type, fn, disabled = false }) => {
   return (
@@ -13,8 +14,10 @@ const TextButton = ({ type, fn, disabled = false }) => {
 const StyledTextButton = styled(Button)`
   font-size: ${theme.fontSizes.M};
   font-weight: bold;
-  background: ${(props) => (props.disabled ? theme.colors.lightGrayBG : theme.colors.green)};
-  color: ${(props) => (props.disabled ? theme.colors.gray : theme.colors.white)};
+  background: ${(props) =>
+    props.disabled ? theme.colors.lightGrayBG : theme.colors.green};
+  color: ${(props) =>
+    props.disabled ? theme.colors.gray : theme.colors.white};
   cursor: ${(props) => (props.disabled ? "not-allowed" : "pointer")};
   width: 440px;
   height: 58px;
