+++ banchan/src/components/utils/variables.js
@@ -1,5 +1,3 @@
-import styled from "styled-components";
-import theme from "../constant";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faSearch,
@@ -21,9 +19,9 @@ const BUTTON_TYPE = {
   SOLDOUT: "일시품절",
 };
 
-const Button = styled.button`
-  all: unset;
-  border-radius: ${theme.borders.radius};
-`;
+const LABEL_TYPE = {
+  EVENT: "이벤트특가",
+  LAUNCH: "런칭특가",
+};
 
-export { BUTTON_TYPE, Button };
+export { BUTTON_TYPE, LABEL_TYPE };
